+++ switch-vision-unifi2mqtt/run.sh
@@ -2,7 +2,7 @@
 set -eu
 umask 077
 
-VERSION="4.0.2"
+VERSION="4.0.3"
 bashio::log.info "Switch Vision UniFi2MQTT v${VERSION} starting."
 
 mkdir -p /share/switch_vision/unifi